+++ ci/environment.yml
@@ -13,3 +13,6 @@ dependencies:
   - pytest-cov
   - scipy
   - xarray
+  # Tools run by pre-commit so they can be in the same conda environment
+  - yapf
+  - docformatter
