--- test/meshfiles/fesom/fesom.clock
@@ -1,2 +1,0 @@
-   85500.000000000000                1        1948
-   86400.000000000000                1        1948
